+++ gamemodes/terrortown/entities/entities/ttt_traitor_check.lua
@@ -32,7 +32,7 @@ function ENT:CountValidPlayers(activator, caller, data)
     local mins = self:LocalToWorld(self:OBBMins())
     local maxs = self:LocalToWorld(self:OBBMaxs())
 
-    local plys = select(2, player.Iterator())
+    local plys = player.GetAll()
     local count = 0
 
     for i = 1, #plys do