+++ gamemodes/terrortown/gamemode/server/sv_shopeditor.lua
@@ -4,7 +4,7 @@
 local net = net
 local sql = sql
 local util = util
-local playerIterator = player.Iterator
+local playerGetAll = player.GetAll
 
 util.AddNetworkString("TTT2UpdateCVar")
 
@@ -95,7 +95,7 @@ function ShopEditor.AddToShopEditor(ply, roleData, equip)
     AddEquipmentToRole(roleData.index, eq)
 
     -- last but not least, notify each player
-    local plys = select(2, playerIterator())
+    local plys = playerGetAll()
     local nick = ply:Nick()
     local rdName = roleData.name
 
@@ -127,7 +127,7 @@ function ShopEditor.RemoveFromShopEditor(ply, roleData, equip)
     RemoveEquipmentFromRole(roleData.index, eq)
 
     -- last but not least, notify each player
-    local plys = select(2, playerIterator())
+    local plys = playerGetAll()
     local nick = ply:Nick()
     local rdName = roleData.name
 
@@ -237,7 +237,7 @@ function ShopEditor.OnChangeWSCVar(subrole, fallback, ply_or_rf)
     if fallback ~= SHOP_UNSET and fallback == rd.name then
         LoadSingleShopEquipment(rd)
 
-        ply_or_rf = ply_or_rf or select(2, playerIterator())
+        ply_or_rf = ply_or_rf or playerGetAll()
 
         if not istable(ply_or_rf) then
             ply_or_rf = { ply_or_rf }