+++ gamemodes/terrortown/gamemode/shared/sh_player_ext.lua
@@ -8,7 +8,6 @@ local table = table
 local IsValid = IsValid
 local hook = hook
 local math = math
-local playerIterator = player.Iterator
 
 -- Distinguish between 3 modes to reset, add or remove equipped items
 EQUIPITEMS_RESET = 0
@@ -1147,7 +1146,7 @@ function plymeta:SetModel(mdlName)
 end
 
 hook.Add("TTTEndRound", "TTTEndRound4TTT2TargetPlayer", function()
-    local plys = select(2, playerIterator())
+    local plys = player.GetAll()
     for i = 1, #plys do
         plys[i].targetPlayer = nil
     end