+++ lua/terrortown/menus/gamemode/gameplay/voiceandvolume.lua
@@ -9,6 +9,21 @@ CLGAMEMODESUBMENU.icon = Material("vgui/ttt/vskin/helpscreen/voiceandvolume")
 function CLGAMEMODESUBMENU:Populate(parent)
     local form = vgui.CreateTTT2Form(parent, "header_voiceandvolume_settings")
 
+    form:MakeHelp({
+        label = "help_voice_activation",
+    })
+
+    form:MakeComboBox({
+        label = "label_voice_activation",
+        convar = "ttt2_voice_activation",
+        choices = util.ComboBoxChoicesFromKeys(
+            VOICE.ActivationModes,
+            "label_voice_activation_mode_",
+            VOICE.cv.activation_mode:GetString()
+        ),
+        OnChange = VOICE.ActivationModeFunc("OnJoin"),
+    })
+
     form:MakeCheckBox({
         label = "label_gameplay_mute",
         convar = "ttt_mute_team_check",
@@ -17,10 +32,16 @@ function CLGAMEMODESUBMENU:Populate(parent)
     form:MakeComboBox({
         label = "label_voice_scaling",
         convar = "ttt2_voice_scaling",
-        choices = VOICE.GetScalingFunctions(),
+        choices = util.ComboBoxChoicesFromKeys(
+            VOICE.ScalingFunctions,
+            "label_voice_scaling_mode_",
+            VOICE.cv.scaling_mode:GetString()
+        ),
         OnChange = function()
-            for _, ply in ipairs(select(2, player.Iterator())) do
-                VOICE.UpdatePlayerVoiceVolume(ply)
+            local plys = player.GetAll()
+
+            for i = 1, #plys do
+                VOICE.UpdatePlayerVoiceVolume(plys[i])
             end
         end,
     })